% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr_fish_doc.R
\docType{data}
\name{vcr_fish}
\alias{vcr_fish}
\title{Virginia Coast Reserve Fish Counts and Lengths}
\format{
An object of class \code{data.frame} with 1720 rows and 17 columns.
}
\source{
{McGlathery, K. 2019. Fish Counts and Lengths in South Bay and Hog Island Bay, Virginia 2012-2018 ver 11. Environmental Data Initiative. https://doi.org/10.6073/pasta/af32b580d4646f4b35d7552d83e9b09b (Accessed 2021-05-03).}
\url{Add link for data package here}
}
\usage{
vcr_fish
}
\description{
Sampling of small fish in South Bay (west of Wreck Island) and Hog Island Bay was conducted using seine nets. Individual fish were identified and their length measured
}
\keyword{datasets}
